tabelasmigrations

+++ backend/database/migrations/2026_06_23_172636_add_id_item_to_movimentacao_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('movimentacao', function (Blueprint $table) {
-            $table->integer('id_item')->nullable()->after('id_lote');
+            $table->unsignedBigInteger('id_item')->nullable()->after('id_lote');
 
             $table->foreign('id_item')
                 ->references('id_item')
